chore: fix android build

+++ android/src/main/java/com/fugood/reactnativevideoplayer/ReactNativeVideoPlayerViewManager.kt
@@ -66,7 +66,7 @@ class ReactNativeVideoPlayerViewManager :
 
   override fun receiveCommand(view: ReactNativeVideoPlayerView, commandId: Int, args: ReadableArray?) {
     when (commandId) {
-      COMMAND_SEEK -> view.seekTo(args?.getFloat(0) ?: 0)
+      COMMAND_SEEK -> view.seekTo((args?.getDouble(0) ?: 0.0).toFloat())
       COMMAND_PLAY -> view.play()
       COMMAND_PAUSE -> view.pause()
       COMMAND_STOP -> view.stop()
